common/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
description = "Shared utilities for the Alquist Insight project"
name = "common"
requires-python = ">=3.10"
version = "0.1.0"

authors = [
    { name = "Ella Brichova", email = "ella.brichova@cvut.cz" },
]

maintainers = [
    { name = "Ella Brichova", email = "ella.brichova@cvut.cz" },
]

dependencies = [
    "bson",
    "cachetools",
    "fastapi",
    "psutil",
    "pydantic>=2.0",
    "pydantic-settings",
    "python-dateutil",
    "starlette",
]

[project.optional-dependencies]
es = ["elasticsearch==8.17.2"]
jwt = ["httpx", "pyjwt[crypto]"]
mongo = ["dnspython", "pymongo"]
openai = ["openai"]
